requirements.txt
PyPDF2 == 3.0.1
langchain == 0.3.25
wikipedia == 1.4.0
cohere == 5.15.0
pinecone == 6.0.0
langchain-community == 0.3.23
langchain-cohere == 0.4.4
fastapi == 0.115.8
uvicorn == 0.34.0
streamlit == 1.42.2
python-dotenv == 1.1.0